Fix serious error in handling of min/max
A serious error in the handling of derivatives of min/max functions
was the cause of my BSIM6 woes.

This error actually impacted non-sensitivity code as well, but
surprisingly not in any way that seemed to matter to any of our
existing models (even the BSIM6) for anything but sensitivity.  We
have been consistently matching CMC test suite results since we were
still using Sacado everywhere --- and that hasn't changed at all with
this fix, either.

At any rate, with this fix, BSIM6 works correctly in sensitivity, so
I'm committing it.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1284