Fix backend floor/ceil, and other mistakes
floor and ceil still had some broken bits left over from their Sacado
legacy.  There were also a few other places where old Sacado code
would have emitted ".val()" appendages onto variables.  These are
gone.

There was also a mistake regarding my attempt to skip assignments and
declarations of output-only variables in sensitivity context (because
they're not used anywhere).  Thing is, it is possible to make a LOCAL
variable (that is neither output nor involved in computing sources,
but which is used in computing OTHER output variables) that references
them on the RHS.

Since there is no easy way to detect that an assignment has a local
variable on the LHS that involves only output variables on the RHS and
is only used later to compute some other output variable, I have opted
instead to fake it out by returning "0.0" from the template for
variables when the variable is one for which we're not emitting
declarations.  We wind up with a bunch of stupid looking "variable =
0.0" assignments when the RHS involves nothing but output variables
but those variables on the LHS are never used and the compiler will
optimize them away.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1284