Restore C++98 compatibility to Xyce/ADMS output
My merge of my bug1284 branch had the unintended consequence of
rendering Xyce uncompilable on RHEL6.  We have dropped support for
RHEL6, but one of the developers still has a VM that is in use
for something important.

The issue was the unnecessary initialization of class member variables
in the header file.  This is illegal in C++98 (though fine in C++11),
and was not really intentional --- it is the side effect of my
forcing *other* variables to be initialized on declaration.

This commit adds logic to prevent this from happening in creation of
the .h file.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1284