Revert "Extend S-parameter device to include short circuit currents"
This reverts commit 8b5922ff421febab63f927eda52c39e4db1e6662.

The commit breaks the existing YLIN tests in parallel. It works
fine in serial. It adds short circuit currents to the S-parameter
model which load to the frequency domain B vector. However, the
frequency domain loading infrastructure in HB is never tested for B
vector in both serial and parallel since no frequency domain device
has a B vector previously. We found that the potential problem is that the B
vector does not address off-processor contributions which can cause
problem for current sources. Before this is fixed, the commit is reverted
so the YLIN tests can run in parallel.