revamping methods for resolving functions & params
I have been working on merging the new expression library into Xyce, to
replace the old one.  I originally designed the new expression with the
concept of a "group", which would provide information from the simulator
to the expression.

For things like solution values, lead currents, etc, the group is well
suited for this, in part b/c the usage is basically the same for those
quantities everywhere in Xyce.

However, I had also designed it to be involved in resolving external
dependencies such as calls to .func and calls to .global_param.
However, in Xyce having the group do this doesn't work as well as it
does for solution variables, so I have been looking at handling these
reseolutions outside of the group.

There are a few reasons why the group (maybe) should not be involved in
expression resolution in Xyce.  They include:
 - The logic for resolving a .func or a .param depends on the specific
circuit context.  If resolving an expression in the parser, then all
these resolutions are done at the context level and it is important for
the resolve code to understand which context we are in.  The only way
for this to work is for the group to understand context, and have
deependencies on the IO package.  I don't want that.

 - expressions in Xyce are mostly stored inside of N_UTL_Param objects.  So,
when searching for a .func by name, you don't get the expression directly, you
instead get back the UTL::Param that owns it.  This is fine, but it would mean
that the group object would need to pull in the full headers for N_UTL_Param
and all the related dependencies.  I don't want that either.

Basically having the group be able to do .func and .global_param
resolution requires pulling in a lot of Xyce headers and also makes
the logic confusing.

Anyway, in this commit, I've created some new functions to
newExpression.  They are  attachFunctionNode and attachParameterNode
and the arguments are the name of the object, and the object itself.
This allows the machinery of attaching the external node to still be
handled inside of newExpression, but for all the responsibility of
finding the external objects to attach belongs to the calling code.  The
group is not involved.

I have added a few tests for these new functions to paramUnitTest.C and
they pass.

For the time being, both types of resolution are supported (both in the
group and out of it).  Eventually I should probably get rid of the group
style of doing it, but that will require many tests to be changed.
So, for the time being they are both still there and I've tried to make
them co-exist without making a mess.