First cut of correct use of .func.
I have revamped the way that expressions handle .funcs.  This only works
with the new expression library, and does not work with the old one.  I
have abandoned the old API and am creating a new one.

This new implementation has so far only been implemented in the circuit
context class.  Other places that use .func (like sensitivities, or the
.print line) have not been updated yet.  That will come later.

I have also abandoned the idea of having the "group" class provide
.func,.param and .global_param resolution.  There is too much
complexity in the circuit hierarchy to do this very cleanly.  So, I will
still use a "group", but the group now will *only* supply things like
solution values, state values, etc.  It will not do the initial setup
work such as parameter resolution.

As I mentioned in my previous commit, I have added two functions to the
newExpression class:

newExpression::attachFunctionNode
newExpression::attachParameterNode

And these basically are used in place of the group.  In theory, the
second one can be used for the .param and .global_params, but this
hasn't happened yet.  The first of these function is now used by the
circuit context, as part of N_IO_CircuitContext::resolveFunctions.  I
have not used the "attachParameterNode" function in the context yet.

In using these from the context, I had to be very careful about what
order the setup functions are called, and I wound up having to add an
optional argument (a vector of strings of function arguments) to the
N_UTL_Expression constructor, to enforce that order for functions.

The function newExpression::resolveExpression is on its way to being
obsolete.  However, I will have to modify the unit tests extensively
before I can remove it.  It is not needed by the Xyce usage of
newExpression.

This commit also removes the experimental "parser" expression group.
I was experimenting with a few ideas for groups.  The context-based
resolutions were tricky to handle and I considered having different
groups in different parser contexts.  The idea did not pan out, so
removing these (now useless) files.