Fix, test big inefficiency in funcOp::dx funciton
For the first time in a very long time I started testing the use case
that motivated the development of this new expression library.  That use
case is one in which you have a really large number of nested .func
calls in the netlist.  In the old expression library, the code went into
an n^2 or operation and never came back.

The old expression library kind of treated .funcs as pre-processor
macros and set them up in the expressions that called them via string
replacements, followed by a re-parse and then followed by a error check
of the syntax tree.  When you have to do this 1000's
of times, the parsing and checking is never done.

The new expression library has been designed to handle .funcs more like
a function in C, in that the function just has a pointer, and to execute
the code in the function you traverse the tree thru that pointer.  In
this approach, the .func is only parsed 1x.

Really early in the development of this library, I tested this very use
case in a simple test harness, and it appeared to scale pretty well to
really large numbers, thus validating the design.   Unfortunately, when
I started setting up unit tests, I never created any unit tests for
really large numbers of nested functions.  That changed today.  In this
commit I am adding two tests:

Double_Parser_NestedFunc_Test:1000nest_no_deriv
Double_Parser_NestedFunc_Test:200nest_with_deriv

In both cases, I am creating the simplest possible function call tree,
where all but one function is a simple pass-thru (i.e. does nothing but
proceed to next function pointer).

I discovered that the "newExpression::evaluateFunction" function executed
fairly efficiently (although not as efficiently as I would like) on 1000
simple nested functions.   However, newExpression::evaluate did not.
In fact, "evaluate" bogged down and did not come back.

I discovered the reason. unlike "evaluateFunction", "evaluate" also
computes the derivatives.  And, the "dx" function in the funcOp class
had an inefficiency in it, forcing it to traverse down the tree a lot
more frequently than it needed to.  In this commit I am adding a very
simple check to skip traversals that are not needed.  (if the thing
being retreived by that traversal is being multiplied by 0.0, then no
point in traversing).

I am sure there are more efficiency improvements to be found, but this
one was really easy to do, and made the motivating use case work with
derivatives.