Add new back-end template file to installed list
"make install" installs ADMS back-end XML files so that one can
compile Verilog-A models.

I added a new XML file as part of bug 788, xyceOutVarsDoc.xml, that
generates tables of output variables defined in the model for
inclusion in the reference guide.

Let's install that one at the same time as we install others.