Add a step delay element to Xyce
This is for bug 1293 on Joseki.

It is intended to provide something like a voltage-controlled voltage
source, but that has an output voltage that is the same as what the
control voltage had one timestep ago.

An important point: At one point in the development of this device I
had tried to use a Master class that pretty much just called the
instance class loadDAEFVector and loadDAEBVector functions.  I was
rewarded for my effort with a device that produced HALF of the voltage
it should have.

I don't know what it was about how I had defined and used the "Master"
class for this device, but throwing it away and using the default
implementation (which does what I had *thought* was all mine did)
simply made this work correctly.

Since I honestly don't care about all the things that using a custom
Master class could possibly give me, I'm happy with leaving it this
way.  Should we ever need what a custom Master class would provide,
we can revisit this choice.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1293