Patch around broken sensitivity code output
So, during my work on bug 1284, I found that a lot of output variable
assignments were being generated in sensitivity context for absolutely
no reason.  These variables are never used in that context, so it is
pointless to compute them.  I figured out how to get ADMS to skip
declaring them or emitting any assignments to compute them.  That was
easy, because these variables are tagged with "output='yes'" and
"insource='no'".

Then came PSP, which not only has tons of output variables, but a few
local variables that have pure output variables on the RHS and which
are used for no purpose other than to be on the RHS of other output
variable assignments.  ADMS does not, however tag these things in any
way except that they're insource='no' (and that isn't enough to know
that they can be skipped).

I kludged up a hack so that if it turns out that output variables that
are never used in sources are used anywhere in sensitivity context,
that I just emit "0.0" for them instead of a variable reference.  "It
doesn't matter --- the use of these variables on the RHS is to compute
variables on the LHS that are never, ever used and will be optimized away."

That, unfortunately, doesn't work if these variables appear in denominators.
Intel 2019 on Windows flags this as a divide by zero error at compile time.
Ooops.

This impacts PSP102, PSP103 (but not JUNCAP200), MEXTRAM, and HICUM
level 0 (but not level 2).

This commit replaces my dumb placeholder of 0.0 with 1.0, which should
work everywhere.  It also fixes up an issue in the template that
handles division so that if, at the time ADMS has finished collapsing
denominators, it finds a denominator of 1.0, it simply emits the
numerator instead of numerator/1.0.

This should fix the compilation error that Rich just discovered on
Windows with Intel 2019.