Make Xyce/ADMS more agressive about skipping assignments
In my original bug 1284 work, I had "optimized" sensitivity code by
skipping computation of output pure variables (variables that are
designated for output, but do not get used in computation of
sources). It did this solely by looking at the LHS of assignments, and
skipping the assignment *AND DECLARATION* of such variables.

That was a problem, because *sometimes* (mostly in mextram and hicum
level 0), there were local variables that were also not used in
sources (the only thing that matters in sensitivity context at this
point), but were not output variables, but whose assignment involved
output variables on the RHS *AND* which were subsequently ONLY used to
calculate other output variables (and therefore these variables were
never actually used in sensitivity context).

My first approach to dealing with this was to carelessly replace all
now-undeclared output variables on the RHS of such assignments with
0.0.  That was dumb, because sometimes these variables were used in
denominators --- the Intel 2019 compiler then catches the division by zero
as a compile time error on Windows.

My second approach (earlier this evening) made the substitution 1.0
instead of 0.0.  That fixed the divide by zero, but still had several
stupid, unnecessary asignments.

This commit adds more aggressive pruning of assignment statements in
sensitivity functions.  We skip an assignment if:
  - the LHS is a pure output variable (never used in sources)
OR:
  - the LHS isn't output, but also isn't in sources
     AND
     - the RHS contains output variables that are not in sources and
       are not input variables (parameters).

This removes a handful of completely pointless assignments in MEXTRAM
and HICUM level 0.  It has no impact on other models.  It removes the
dumb divide by zero problem, but doesn't do so by introducing
pointless computations involving lots of 1.0 which are then assigned
to variables that are never, ever used, either.

Note that I don't, in fact, skip *declaration* of these useless
variables (unlike the skipping of declaration of pure output
variables.  I simply don't ever assign to them after they're
initialized to zero.

Note also that this change doesn't eliminate *ALL* pointless
assignments.  In MEXTRAM, there are in fact variables whose
assignments have, on the RHS, only variables whose assignments we just
skipped, but which don't themselves have RHS's that fit our "skip this
assignment" test.  These variables are also useless, but we cant
detect that.  They wind up with expressions on the RHS that involve
only variables that are always zero (because we never assign to them
after initialization by zero).  Solving this problem completely will
involve a really heavy duty recursive operation, and is not worth the
trouble.  I'm just trying to avoid generating useless expressions that
can be detected easily, and to not replace them with clearly illegal
constructs that will cause problems at compile time, or crash at run
time.  The current commit seems a reasonable compromise.

Again, this stuff only applies in sensitivity, where output variables
are all useless and only contributions to sources matter.  It has no
impact on normal, non-sensitivity code.