Fixes parallel load of frequency domain B vector
The initial B vector load in the frequency domain was not expecting
communication to perform the load.  With the YLIN device, it will
be necessary for contributions to be sent and received from other
processors.  This is necessitated by the fact that the frequency
domain vectors are based on the time-domain loading, which will
perform the communication before returning the resulting data.
The frequency-domain linear algebra infrastructure does not perform
the communication at this time.  This has been added to the B vector
loading for HB.