Checkpoint commit on the way to global expr group
I have started working on implementing a "main" xyce group in earnest.

This commit does a handful of things.

(1) it comments out the Util::Expression::evaluate and
Util::Expression::evaluateFunction functions that require a values
vector passed into them as function arguments.  These will not make
sense for the "group" design.

(2) commenting out all usage of these evaluate functions, and replaced
with calls to the "non-value" versions.

(3) Also commenting out the "set_values" functions of Util::Expression.
In general, I don't want the expression library to rely upon passing
values in directly, and I don't want those values to be in a single,
double-precision array anyway.

(4) this commit adds a few "set" functions to the Util::Expression
class, but I'm not sure if I will keep them.  They are better than the
stuff I commented out, but still not great.

(5) In two places in the ANP::Transient class, there were expressions
that were allocated only for the express purpose of setting the
"accepted time".  In the old expression library this accepted time is a
static varbiable, so any expression can set it, even one that isn't
used.   In the "group" model these calls (and expressions) won't be
needed so they are commented out.

(6) work on the "mainXyceExpressionGroup" class.  This isn't ready for
prime time.  But the goal of this group is for it to be the provider of
Xyce state information to expressons.  The intent is for there to only
be one of these, and for it to be passed into all expressions when then
are constructed.    I have started fleshing out some of the functions
but they aren't used yet.

This code will compile but will not really work.