Add multiplicity to BSIM6
The BSIMs generally aren't coded to have explicit multiplicity
factors, and rely on some old Verilog-A LRM that claimed it should be
implicit.  Xyce does NOT have implicit multiplicity and requires each
device to code it explicitly.

CMC has some standard macros to use to turn on or off explicit
multiplicity, because some simulators are like Xyce and don't do it
implicitly.  BSIM doesn't use them (in some models they have the macro
defined but not used, in others they simply claim that the LRM 2.2
says it's implicit and leave it out).

This commit hacks it into BSIM6, as I had to do for BSIM-CMG before
it, and am doing to B4SOI soon.

Passes CMC test suite for all "m" variant tests.

As an aside, I am also removing the Xyce-specific hacks I had put in
when Xyce/ADMS didn't support ddx(a+b,v(c)).  I had removed those
hacks in other models last week when I fixed that issue, but neglected
to revert it here.

Also updated README_Xyce to reflect the current set of differences
from the original code.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1296