Modify BSIM SOI 4.6.1 to work in Xyce
A small handful of changes made to the Verilog-A file allows this to
process in Xyce and pass most (but not all) of the tests in the CMC
test suite.

The failing tests in the CMC test suite are all AC tests, and are
failing due to tiny numbers (much smaller than machine epsilon) in the
"standard" results file.  The CMC test script uses the maximum
relative error as its error norm, and these tiny numbers are
essentially just round-off error.  Xyce and the gold standard get
different round off on one or two entries, and this makes the test
appear to be failing the test.  Careful examination of the actual
output shows that the nature of this "failure" is meaningless.

This commit modifies the Verilog-A file, adds a Makefile, and updates
our ADMS-related utility scripts to regenerate the BSIM SOI device as
needed.

Differences between the Xyce-compatible version and original BSIM
device are documented in code/README_Xyce.

Xyce-compatible versions of the tests in the "benchmark_test"
directory, and gnuplot-compatible copies of the HSPICE output for the
reference runs are also included here.

https://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=203