Convert to all expressions using a global group
This commit touches nearly every expression usage in Xyce.  There are a
few exceptions.

(1) in N_UTL_Param, there are 2 expression constructions that for now I
have just commented out.  I did this b/c I don't think the Util::Param
classes necessarily should have direct access to this group.  Also,
giving them this access felt like a real pain.  So that is commented out
for now; I don't think it affects much and hopefully I will find a
workaround.

(2) in a few other places expressions are *just* used to help parse and
are never meant to be evaluated.  For a couple of those cases (forget
where) I just created a base-class group locally (that did nothing and
could resolve nothing) and used it instead.

Otherwise, they all use the global group now.

The global group doesn't do very much  yet, so many things are broken.
Also, a bunch of leftover code from my previous attempt at getting the
new expression library to work in Xyce is still there (particularly in
N_UTL_Expression.C) and now it is in the way causing problems.  I
removed some of this code but not all yet.

But this is a reaonable checkpoint for a commit.  Despite touching every
expression, the code compiles and can run a few things with expressions
correctly.