Move File Readers Used for -remeasure and (now) YLIN
This commit moves the files, and classes, for re-reading Xyce
output files from the Measure namespace to the IO namespace.
This is being done because those classes will now be used for
both -remeasure and the YLIN device.  In addition, the OutputFileBase
class was already in the IO namespace.  So, this re-factor makes
those class/file locations more consistent.

SON Bug 1295
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1295