Fix up Expression::replace_var and other related
Today I worked on debugging all the .sens tests for my newExpression
implementation.    I did this previously for my attempts to use it with
the old API, and now I am doing it again.  I found a number of problems.

Problems fixed in this commit:

- get_names(STRING) needed to only return unresolved strings.  In my
newExpression library, "string" roughly means "parameter name".   I had
mistakenly written the code to return the names of all parameters,
resolved or not.  This has been fixed.

- Expression::replace_var was not implemented.  I had mistakenly thought
its purpose was much smaller than it actually is.  It is a good
candidate to perform an "newExpression::attachParameter"  function call
underneath the covers.  It is for (global) parameters that are not constants, and
are not variables, but are instead expressions.

- Some of the error checking in the CircuitContext class for funcs was
not necessary anymore, and was causing problems.  After the function is
resolved, the old code was re-parsing it to double check that the
resolution was correct.  And by "correct" the only remaining unresolved
strings were function arguments.  The problem for newExpression with
this error check is that it performed it on a new expression, which was
built from the string of the old expression.  As I do not rely on string
substitutions to process params or funcs, this string was 100% the same
as before the resolution.  So, all the "unresolved" strings would still
seem to be present.     But, my new code had other error checks in
place, and so this error check (even if it had been valid) was a waste
of time.  Commented out for now.

I also fixed a number of problems in the mainXyceExpressionGroup class,
but I can't remember if that was in the previous commit or this one.
Mainly, I had to use slightly different function calls to get things
like time, freq, etc.  I had to go thru the device manager rather than
the analysis manager, because the analysis manager's accessors assumed
things like the step error control class were already allocated.  At the
earliest calls that the group class had to make, this was not the case.
The device version of this function call was much safer.