Support Upper and Lower Format for TS2 Input for YLIN
This commit allows the [Network Data] section of the Touchstone2
formatted input file for a YLIN device to also use "Upper" or
"Lower" as the [Matrix Format].  The default format, if the optional
[Matrix Format] line is omitted, is "Full".  Note that those format
labels are case-sensitive.

If the model parameter ISC_FD=true then only Full format is
supported for now.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215