First cut at supporting Xyce complex outputs
Make it possible for new expression parser to recognize the following
operators:

VR, VI, VM, VP, IR, II, IM, IP

these are all used for complex numbered output on the .print line.

To support this, I did something very similar to  how I supported things
like V(n1,n2), in which the output should be V(n1)-V(n2).  The easiest
way to do this was to apply an extra binaryPlusOp as the operator
returned for the voltage op.

Similarly, in this case, for things like VR, I don't return a voltageOp
directly, but instead return a realOp that used that voltageOp as an
input.

This doesn't yet mesh with the way that Xyce provides these types of
numbers.  Still have to figure this out.

Also, I have added some unit tests for this, but not exhaustive yet.  So
far, only tests in parserUnitTest.C , which is primarily a
double-precision oriented set of tests.  I still need to create tests in
the complex numbered unit tests (complexParserUnitTest.C).  The complex
tests are way out of date, so they need much more updates than just
this.

Finally, I also modified the lexing of V(node), etc, to accomodate
whitespace between the V and the parens.