Update complex tests to be closer to double tests
This was overdue, and there is still work to be done.

I have been working on complex operators today, such as VR, VI, VM, VP.
These are most pertinent to using std::complex.  However, I had not been
testing with std::complex recently, and the complex unit tests were a
few weeks out of compliance.  This commit fixes that issue.

It does not add any tests for VR, etc.  That will come later.