Make YLIN Handling of Touchstone Input Case-Insensitive
This commit improve the handling of the Touchstone 2 input file
for the YLIN device.  Previously, the handling of the meta-data
tags in those files, such as [Network Data], was case-sensitive.
That error is now fixed, as each line of the Touchstone 2 input file
is upper-cased before further processing of it occurs.

Enums were also added for ParameterType, MatrixFormat and DataFormat.
These replace string compares in the function Model::readTouchStoneFile.
This further reduces opportunities for case-sensitivity to creep back
into the Touchstone 2 file parsing.

SON Bug 1215
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1215