More work on complex number expressions, cleanup
This commit contains a lot of changes.

Yesterday, I started working on supporting operators like VR, VI, VM,
VP, which are for complex numbererd outputs on the .PRINT line (and in
some expressions).

These (and others I haven't gotten around to yet) had been supported in
the old expression library, but I had put off supporting them in the new
expression library.  They are in there now, and mostly set up, but
getting them to talk properly to Xyce is still a work in progress.

As part of this work, I felt it was necessary to go back and start
testing the expression library with complex numbers again.  So I spent
the rest of the day yesterday getting the complexParserUnitTest program
running again, and added tests to it for things like VR, VI.
Thankfully, it seems to still work pretty well.

I have a pipe dream that I can push templating all the way thru the new
expression library.  Unfortunately, working with Bison makes this hard
to do, and it might be impossible.  So, for the time being, the main
data type used by the expression library code must be known at compile
time.  There is a typedef for "usedType" that is used everywhere.  By
default "usedType" has been "double" but it can also be set (from the
compile line) as "std::complex<double>".

For the time being, I don't think my pipe dream is going to happen.  If
it could happen, then it would be possible to have a single, templated
expression library that could be invoked as either "double" or
"std::complex<double>" at runtime.  That way, if you are running in
transient, you would allocated a "double" expression library, and if you
are running a frequency domain analysis (AC or HB) you would instead
invoke std::complex library.

So, instead, I need a way to support both double and std::complex
without run time template support.

My current plan is to make Xyce use std::complex inside the expression
library, but then have accessors that support both double and
std::complex.  So, if evaluating an expression that is purely double, it
will still use std::complex and the imaginary entry will just be zero,
and the real componenet is what is output to the rest of Xyce.

I have a first cut of this working now.  The code compiles and runs with
either double or std::complex<double> as the expression type.  And, the
same tests pass and fail.  So I have not broken anything by using
std::complex.

I had to hack a fair amount on the mainExpressionGroup to support
complex, and there is a lot more work to be done there.

Also as part of this commit, I have done a lot of code cleanup in the
N_UTL_Expression class.  I have removed all possibility of using the old
expression library.  There is no longer an expPtr present in the class.
There is no longer a boolean variable that determines which library is
in use.  This allowed a lot of if-statements to go away.

I have also removed a lot of functions that were not being used, and
will (hopefully) never need to be used again.  I have also fixed up a
lot of the comments and function headers.