Fix for spice_pulse breakpoints
Today I started debugging the use of spice source expressions in the new
expression library, when used in Xyce.  I had previously created various
unit tests for this capability but those tests were kind of in the
category of "a good start, but not nearly enough".

One thing I was not testing at all in the unit tests was the
getBreakPoints  function.  My plan all along has been to completely do
away with the old expression library's style of breakpointing, which
determins, in a very weird ad-hoc manner, the nearest breakpoint an just
returns that one.  And it did it very generically; it didn't matter what
the function was, and it relied on a quasi Newton calculation to find
it.

For new expressions, I chose instead to just re-implement the original
SPICE independent sources as faithfully as I could.  ie, copy the code
directly out of the Vsrc, and then hack it to work with operators.  That
way, each spice_source could have its own unique breakpoint function (if
necessary) and the code would just call it.  I did about 90% of the work
to set this up, but didn't quite finish it.

This finishes it, at least on the newExpression side, for the
spice_pulse operator, and for the general newExpression structure.  I
have also created a new unit test for it.