Fix alias node map usage in main expr group
I threw the use of the alias node map (which is created in the netlist
import tool, but used in many places in Xyce), but had not tested it.

I looked more closely at it today and realized that I had goofed and
created an invalid pointer.  Fixed now.

Before this fix, the Certification_Tests/BUG_1962/sens.cir test failed
with a seg fault.  Now it passes.  This may have been the only
sensitivity test that depended on it.