Begin lead current and internal device var support
This is a checkpoint commit.  I have begun to support internal device
variables, which are specified by N() and also lead currents.

I have been copying and modifying some of the code out of the OpBuilders
class to support some of this.  I can't use the Op code directly as it
is designed with a number of assumptions that are not operative for the
new expression library.  One big assumption is that it will handle all
of the complex operators (like VR, NR, etc) itself.  In the new
expression library, these are parsed and handled by the expression
library, which is a cleaner way to do it.

This is not done yet.  Just wanted to checkpoint my work.  This commit
compiles, and a handful of tests now pass that previously failed.