More support for N() operators, and tests
I had previously forgotten that the N() operator doesn't just apply to
internal, locally defined variables like "gm" in a mosfet.  It also
applies to internal solution vars as well, that can be in the solution
vector.

As such, support for the N() operator needs to support all the complex
numbered variants: NR, NI, NM, NP, NDB.  This commit adds support for
them, except NDB which is only 1/2 done and not tested.  The other 4
have tests now.  Also, have added VM and VP tests as well.