Created an lcov test for the complex unit tests
Follows the same pattern as the original lcov tests.  Unfortunately, as
written I can't browse the results from both at the same time, as the
names of the files that contain lcov results are based on the names of
the sources files, so if I run double precision lcov followed by
std::complex<double> lcov, the second run overwrites the first.

But it is still useful.  Not surprisingly, the coverage is worse for the
complex tests, as I haven't been 100% diligent in updating them.  So,
they lag behind the double tests by a little bit.