Duplicate Max Measure Files as New Extrema Class Files
This commit copies the files for the Max measure class to the
files for the Extrema class.  That new class will be used as a
base class for the Max, Min and PeakToPeak classes.  This duplication
should preserve the history/blame of the Max class as the history
for the new Extrema class.

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285