Refactor MAX, MIN and PP Measure Code
This commit refactors the code for the Max, Min and PeakToPeak
classes to put all of their common functions into the new Extrema
class that is derived from the Base class.  It also simplifies the
Max, Min and PeakToPeak classes so that a common set of functions
are called (in each of those classes) by the UpdateTran, UpdateDC
and UpdateAC functions in the Extrema class.  This should eliminate
most of the duplicate code in those three classes.

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285