Inductor IC data now passed to Mutual Ind Const.
The initial condition data from the component inductors is now
passed in the instance block to the constructor of the linear
mutual inductor.  So the needed IC data is now at least getting
to the device instance constructor.  Next step will be getting
the device to use it poperly.

Issue: Mutual Inductors ignore instance IC values.