Duplicate Average Measure Files as New Stats Class Files
This commit copies the files for the Average measure class to the
files for the Stats class.  That new class will be used as a
base class for the Average, IntegralEvaluation and RMS classes.
This duplication should preserve the history/blame of the Average
class as the history for the new Stats class.

SON Bug 1285
http://joseki-vm.sandia.gov/bugzilla/show_bug.cgi?id=1285