Get mutual inductors access to instance IC data
Both linear and nonlinear mutual inductors now get the initial
conditiond data from the component inductor instances. The data
is part of the instance block passed into the constructor.
Unfortunately because the default parameters are added to the
component indcutors and the converted to strings when they are
bundled up into mutual inductors, it's unknown to the mutual
inductor constructor if the initial condition value was given
or if it was the default.  So for now, if the initial condition
value passed in is the default value then the constructor assumes
it was not given.

Issue Mutual Inductor IC bug