Add conditional so empty strings are not parsed
Occasionally Xyce will create an expression object that has an empty
expression string.  The code was written in such a way that even if
string was empty, the parse would get called, and then have nothing to
do.  Bison, in this case, throws an error.

But it wasn't really a valid error and just caused confusion.

So, adding some logic so that empty strings are not parsed.