Support for variable # of args in SFFM spice src
Same type of fix that I appled to the spice_pulse, spice_sin, spice_exp
in previous commits.

Also, there is now one unit tests for a spice source with less than the
full set of arguments.  I had added it to test this capability, then
modified it to pass (by having all the arguments) and now have changed
it back to the original test with 5 (instead of 6) args.