Modified getCurrentVal function to avoid seg fault
I added a lot of code to the various "get" functions in the
mainXyceExpressionGroup a few days ago.  Much of this code is pretty raw
and error prone.  With this commit, I am commenting out some of the
error prone code, which was mostly causing problems (ie seg faults) if
it actually got called.  Thankfully it didn't get called much, but it
did cause a number of tests to fail that should have passed.